requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
google-auth==2.40.3
google-genai==0.3.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
migrate==0.3.8
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==14.2
